shares/util: Move RunCommand to shared/subprocess
